backpack file

+++ WebApi/WebApi/Models/Inventory.cs
@@ -7,7 +7,7 @@ namespace WebApi.Models
         [Key]
         public int id { get; set; }
         public string ItemType { get; set; }
-        public long CategoryID { get; set; }
+        public int CategoryID { get; set; }
         public string ItemName { get; set; }
         public decimal SalesPrice { get; set; }
         public string GSTTaxRate { get; set; }
@@ -18,6 +18,6 @@ namespace WebApi.Models
         public long HSNCode { get; set; }
         public string Description { get; set; }
         public string Image { get; set; }
-        public long BusinessId { get; set; }
+        public int BusinessId { get; set; }
     }
 }
